+++ autoware_utils_uuid/README.md
@@ -1,0 +1,11 @@
+# autoware_utils_uuid
+
+## Overview
+
+The **autoware_utils** library is a comprehensive toolkit designed to facilitate the development of autonomous driving applications.
+This package provides essential utilities for UUID.
+It is extensively used in the Autoware project to handle common tasks such as generation and conversion of UUID.
+
+## Design
+
+- **`uuid_helper.hpp`**: Utilities for generating and managing UUIDs.
